Once again only write pkginfo.g files when the hash is new
This was modified when we merged the tools repository. But it was only
needed due to a minor mistake in the conversion of 'add_sha256_to_json'.
The latter was fixed in d76a56a867267405e2d4903cc7c1204eba4517c8, so it
should be safe to restore this code, too