tools: refactor gap_exec
Directly pipe input into the GAP process, instead of doing it indirectly
via `echo`. This allows us to completely avoid shell quoting issues.

Moreover, capture the output of the GAP process, so that in the future
we can e.g. let GAP print JSON to stdout and parse that in Python.