Fix crashes in digraphs and semigroups (hopefully)
This patch is an attempt to resolve random crashes we have been seeing in CI
whenever digraphs and/or semigroups are loaded. The theory is that this is
caused by their use of `-march=native`, combined with the fact that we use
`ccache`, yet the results run on different VMs. So perhaps one day a binary is
compiled on one computer, and put into the cache; then another day the same
file is compiled again but on a different computer, and the compilation result
is retrieved from the cache. But since the computer has a different CPU, the
binary does not actually run there due to its use of `-march=native`.

Time will tell if this helps or not...