Fix bug that caused failures in PackageManager tests (#1087)
Don't delete build/ffdata.h and build/version.h as PackageManager
needs them to build GAP packages with kernel extensions.

This was a latent bug since 2022, but it was obscured by
PackageManager and GAP library behaviour (e.g. installing  e.g.
`io` was skipped because it was already installed and workable
anyway). Both are stricter now (which is good), and so this bug
in the PackageDistro setup was revealed.